♻️ app: refactored pages/Dash/Dash.tsx
